# Conflicts:
#	Engine_Master_UPC/3rdParty/DirectXTK12/x64/Debug/DirectXTK12.log
#	Engine_Master_UPC/3rdParty/DirectXTK12/x64/Debug/DirectXTK12.vcxproj.FileListAbsolute.txt
#	Engine_Master_UPC/Engine.vcxproj.filters
#	Engine_OUT/imgui.ini

+++ Engine_Master_UPC/DebugDrawPass.cpp
@@ -583,8 +583,8 @@ void DebugDrawPass::apply(ID3D12GraphicsCommandList4* commandList)
     implementation->mvpMatrix = *m_view * *m_projection;
     implementation->commandList = commandList;
 
-    implementation->width = m_viewport->Width;
-    implementation->height = m_viewport->Height;
+    implementation->width = (uint32_t) m_viewport->Width;
+    implementation->height = (uint32_t) m_viewport->Height;
 
     dd::flush();
